+++ src/interfaces/IOFTV2.sol
@@ -8,7 +8,6 @@ import "./ICommonOFT.sol";
  * @dev Interface of the IOFT core standard
  */
 interface IOFTV2 is ICommonOFT {
-
     /**
      * @dev send `_amount` amount of token to (`_dstChainId`, `_toAddress`) from `_from`
      * `_from` the owner of token
@@ -19,7 +18,17 @@ interface IOFTV2 is ICommonOFT {
      * `_zroPaymentAddress` set to address(0x0) if not paying in ZRO (LayerZero Token)
      * `_adapterParams` is a flexible bytes array to indicate messaging adapter services
      */
-    function sendFrom(address _from, uint16 _dstChainId, bytes32 _toAddress, uint _amount, LzCallParams calldata _callParams) external payable;
+    function sendFrom(address _from, uint16 _dstChainId, bytes32 _toAddress, uint256 _amount, LzCallParams calldata _callParams) external payable;
 
-    function sendAndCall(address _from, uint16 _dstChainId, bytes32 _toAddress, uint _amount, bytes calldata _payload, uint64 _dstGasForCall, LzCallParams calldata _callParams) external payable;
+    function sendAndCall(
+        address _from,
+        uint16 _dstChainId,
+        bytes32 _toAddress,
+        uint256 _amount,
+        bytes calldata _payload,
+        uint64 _dstGasForCall,
+        LzCallParams calldata _callParams
+    )
+        external
+        payable;
 }
